feat: 完成P2规划查询优化器RBO规则接线和分布式DML双写
主要变更：
1. P2-1 查询优化器：将5条RBO规则接线到生产查询路径
   - PredicatePushdown（谓词下推）
   - ProjectionPruning（投影裁剪）
   - SubqueryFlattening（子查询展平）
   - IndexSelection（索引选择）
   - CommonSubexpressionElimination（公共子表达式消除）

2. P2-2 分布式能力：完善DML分布式双写
   - 新增 dist_dual_delete 方法处理DELETE操作分布式删除
   - UPDATE/DELETE路径接入分布式双写
   - DML完整路径通过Raft propose → apply同步到分布式KV

3. 协议层修复（P0-FN-TYPE等）：函数返回类型保真

4. 存储层P0问题修复：
   - WAL回放实际恢复数据
   - 快照加载失败拒绝启动（--force-empty）
   - B-Tree集成BufferPool实现磁盘持久化

5. 清理：更新 .gitignore 排除临时测试脚本、日志、Python虚拟环境等

测试验证：400项测试全部通过（含1000条DML序列差分比对0 mismatches）

+++ crates/szrsql-storage/src/btree_bench.rs
@@ -162,7 +162,7 @@ fn m1_bench_point_query_latency() {
     let mut bt = BTree::with_default_order();
     let build_start = Instant::now();
     for (idx, &k) in unique_keys.iter().enumerate() {
-        bt.insert(encode_u64_key(k), (idx % 65536) as u16)
+        bt.insert(encode_u64_key(k), (idx % 65536) as u32)
             .expect("insert should not fail");
     }
     let build_elapsed = build_start.elapsed();
@@ -351,7 +351,7 @@ fn m1_bench_insert_throughput() {
     let rand_start = Instant::now();
     for (idx, &k) in keys.iter().enumerate() {
         bt_rand
-            .insert(encode_u64_key(k), (idx % 65536) as u16)
+            .insert(encode_u64_key(k), (idx % 65536) as u32)
             .expect("insert");
     }
     let rand_elapsed = rand_start.elapsed();
@@ -364,7 +364,7 @@ fn m1_bench_insert_throughput() {
     let sorted_start = Instant::now();
     for (idx, &k) in sorted_keys.iter().enumerate() {
         bt_sorted
-            .insert(encode_u64_key(k), (idx % 65536) as u16)
+            .insert(encode_u64_key(k), (idx % 65536) as u32)
             .expect("insert");
     }
     let sorted_elapsed = sorted_start.elapsed();
