+++ crates/szrsql-storage/src/btree_sonnetdb_defects.rs
@@ -108,7 +108,7 @@ fn defect_01_concurrent_split_no_data_loss_basic() {
             for (idx, &k) in keys_vec.iter().enumerate() {
                 let mut guard = bt.lock().unwrap();
                 guard
-                    .insert(encode_u64_key(k), (idx % 65536) as u16)
+                    .insert(encode_u64_key(k), (idx % 65536) as u32)
                     .expect("insert should not fail");
             }
             keys_vec
@@ -172,7 +172,7 @@ fn defect_01_concurrent_split_no_data_loss_minimal_order() {
             for (idx, &k) in keys_vec.iter().enumerate() {
                 let mut guard = bt.lock().unwrap();
                 guard
-                    .insert(encode_u64_key(k), (idx % 65536) as u16)
+                    .insert(encode_u64_key(k), (idx % 65536) as u32)
                     .expect("insert");
             }
             keys_vec
@@ -225,7 +225,7 @@ fn defect_02_null_key_index_consistency_basic() {
 
     // 插入非空 key 混合
     for v in 1..=100i64 {
-        bt.insert(encode_i64_key(v), v as u16).expect("insert");
+        bt.insert(encode_i64_key(v), v as u32).expect("insert");
     }
 
     // 空 key 仍应可搜索
@@ -285,7 +285,7 @@ fn defect_02_null_key_index_consistency_upsert_and_split() {
     let mut bt = BTree::new(order);
 
     // 多次插入空 key（upsert 语义：应只保留最后一个 tuple_id）
-    for tid in 1..=10u16 {
+    for tid in 1..=10u32 {
         bt.insert(Vec::new(), tid).expect("insert empty key");
     }
     assert_eq!(
@@ -296,7 +296,7 @@ fn defect_02_null_key_index_consistency_upsert_and_split() {
 
     // 插入大量非空 key 触发分裂
     for v in 1..=200i64 {
-        bt.insert(encode_i64_key(v), v as u16).expect("insert");
+        bt.insert(encode_i64_key(v), v as u32).expect("insert");
     }
 
     // 空 key 仍应是 tuple_id=10
@@ -339,7 +339,7 @@ fn defect_03_unique_index_duplication_basic() {
     let key = encode_i64_key(12345);
 
     // 串行插入同一 key 100 次，每次不同 tuple_id
-    for tid in 1..=100u16 {
+    for tid in 1..=100u32 {
         bt.insert(key.clone(), tid).expect("insert");
     }
 
@@ -358,7 +358,7 @@ fn defect_03_unique_index_duplication_basic() {
 
     // 混合插入其他 key 后再 upsert 原 key
     for v in 1..=50i64 {
-        bt.insert(encode_i64_key(v), v as u16).expect("insert");
+        bt.insert(encode_i64_key(v), v as u32).expect("insert");
     }
     bt.insert(key.clone(), 999).expect("upsert");
     assert_eq!(
@@ -389,7 +389,7 @@ fn defect_03_unique_index_duplication_concurrent_upsert() {
         let key_clone = key.clone();
         handles.push(thread::spawn(move || {
             for i in 0..per_thread {
-                let tuple_id = ((tid * per_thread + i) % 65536) as u16;
+                let tuple_id = ((tid * per_thread + i) % 65536) as u32;
                 let mut guard = bt.lock().unwrap();
                 guard.insert(key_clone.clone(), tuple_id).expect("insert");
             }
@@ -433,7 +433,7 @@ fn defect_04_stale_separator_search_basic() {
 
     // 插入 1..=200，触发多次分裂，产生多层 internal 节点
     for v in 1..=200i64 {
-        bt.insert(encode_i64_key(v), v as u16).expect("insert");
+        bt.insert(encode_i64_key(v), v as u32).expect("insert");
     }
     assert!(
         bt.height() >= 2,
@@ -481,7 +481,7 @@ fn defect_04_stale_separator_search_all_deleted() {
 
     // 插入 1..=100
     for v in 1..=100i64 {
-        bt.insert(encode_i64_key(v), v as u16).expect("insert");
+        bt.insert(encode_i64_key(v), v as u32).expect("insert");
     }
 
     // 删除全部
@@ -505,12 +505,12 @@ fn defect_04_stale_separator_search_all_deleted() {
 
     // 重新插入应正常工作
     for v in 1..=50i64 {
-        bt.insert(encode_i64_key(v), v as u16).expect("reinsert");
+        bt.insert(encode_i64_key(v), v as u32).expect("reinsert");
     }
     for v in 1..=50i64 {
         assert_eq!(
             bt.search(&encode_i64_key(v)).expect("search"),
-            Some(v as u16),
+            Some(v as u32),
             "reinserted key {} not found",
             v
         );
@@ -539,7 +539,7 @@ fn defect_05_concurrent_mixed_ops_invariants() {
     {
         let mut guard = bt.lock().unwrap();
         for v in 0..1000i64 {
-            guard.insert(encode_i64_key(v), v as u16).expect("prefill");
+            guard.insert(encode_i64_key(v), v as u32).expect("prefill");
         }
     }
 
@@ -557,7 +557,7 @@ fn defect_05_concurrent_mixed_ops_invariants() {
                     0 => {
                         // insert
                         guard
-                            .insert(encode_i64_key(v), (v % 65536) as u16)
+                            .insert(encode_i64_key(v), (v % 65536) as u32)
                             .expect("insert");
                         local_inserted.insert(v);
                     }
@@ -622,7 +622,7 @@ fn defect_05_concurrent_mixed_ops_insert_delete_same_keys() {
             for v in 0..key_count {
                 bt_a.lock()
                     .unwrap()
-                    .insert(encode_i64_key(v), ((round * key_count + v) % 65536) as u16)
+                    .insert(encode_i64_key(v), ((round * key_count + v) % 65536) as u32)
                     .expect("insert");
             }
         }
@@ -677,9 +677,9 @@ fn defect_05_concurrent_mixed_ops_insert_delete_same_keys() {
 fn defect_06_bulk_load_equivalence_basic() {
     // 生成 5000 个已排序的 i64 key
     let keys: Vec<i64> = (1..=5_000).collect();
-    let items: Vec<(Vec<u8>, u16)> = keys
+    let items: Vec<(Vec<u8>, u32)> = keys
         .iter()
-        .map(|&v| (encode_i64_key(v), v as u16))
+        .map(|&v| (encode_i64_key(v), v as u32))
         .collect();
 
     // 树 A: 逐条 insert
@@ -718,7 +718,7 @@ fn defect_06_bulk_load_equivalence_basic() {
 fn defect_06_bulk_load_equivalence_edge_cases() {
     // 边界 1: bulk_load 空输入应返回 Err(BulkLoadEmpty)
     let mut bt_empty = BTree::with_default_order();
-    let empty: Vec<(Vec<u8>, u16)> = Vec::new();
+    let empty: Vec<(Vec<u8>, u32)> = Vec::new();
     let result = bt_empty.bulk_load(empty);
     assert!(
         matches!(result, Err(BTreeError::BulkLoadEmpty)),
@@ -727,7 +727,7 @@ fn defect_06_bulk_load_equivalence_edge_cases() {
 
     // 边界 2: bulk_load 单元素
     let mut bt_one = BTree::with_default_order();
-    let single = vec![(encode_i64_key(42), 42u16)];
+    let single = vec![(encode_i64_key(42), 42u32)];
     bt_one.bulk_load(single).expect("bulk_load single");
     assert_eq!(
         bt_one.search(&encode_i64_key(42)).expect("search"),
@@ -737,13 +737,13 @@ fn defect_06_bulk_load_equivalence_edge_cases() {
 
     // 边界 3: bulk_load 后再 insert 应正常工作
     let mut bt_mixed = BTree::with_default_order();
-    let initial: Vec<(Vec<u8>, u16)> = (1..=100).map(|v| (encode_i64_key(v), v as u16)).collect();
+    let initial: Vec<(Vec<u8>, u32)> = (1..=100).map(|v| (encode_i64_key(v), v as u32)).collect();
     bt_mixed.bulk_load(initial).expect("bulk_load");
 
     // 再 insert 101..=200
     for v in 101..=200i64 {
         bt_mixed
-            .insert(encode_i64_key(v), v as u16)
+            .insert(encode_i64_key(v), v as u32)
             .expect("insert after bulk_load");
     }
 
@@ -751,7 +751,7 @@ fn defect_06_bulk_load_equivalence_edge_cases() {
     for v in 1..=200i64 {
         assert_eq!(
             bt_mixed.search(&encode_i64_key(v)).expect("search"),
-            Some(v as u16),
+            Some(v as u32),
             "key {} not found after bulk_load + insert",
             v
         );
@@ -761,8 +761,8 @@ fn defect_06_bulk_load_equivalence_edge_cases() {
     // 边界 4: bulk_load 未排序输入应返回 Err(BulkLoadNotSorted)
     let mut bt_unsorted = BTree::with_default_order();
     let unsorted = vec![
-        (encode_i64_key(10), 10u16),
-        (encode_i64_key(5), 5u16), // 乱序
+        (encode_i64_key(10), 10u32),
+        (encode_i64_key(5), 5u32), // 乱序
     ];
     let result = bt_unsorted.bulk_load(unsorted);
     assert!(
@@ -790,7 +790,7 @@ fn m1_milestone_all_defects_summary_large_scale() {
     assert_eq!(keys.len(), total);
 
     for (idx, &v) in keys.iter().enumerate() {
-        bt.insert(encode_i64_key(v), (idx % 65536) as u16)
+        bt.insert(encode_i64_key(v), (idx % 65536) as u32)
             .expect("insert");
     }
